+++ src/asyncEvent.ts
@@ -1,34 +1,30 @@
 import { BaseEvent, EventOptions, Filter, Handler } from './baseEvent';
 import { AggregateError } from './aggregateError';
 
-export type AsyncHandler<TArg, TSender> = Handler<TArg, TSender, Promise<void>>;
+export type AsyncHandler<TArg> = Handler<TArg, Promise<void>>;
 
-export class AsyncEvent<TArg, TSender = void> extends BaseEvent<
-  TArg,
-  TSender,
-  Promise<unknown[]>
-> {
-  constructor(options?: EventOptions<TArg, TSender, Promise<unknown[]>>) {
+export class AsyncEvent<TArg> extends BaseEvent<TArg, Promise<unknown[]>> {
+  constructor(options?: EventOptions<TArg, Promise<unknown[]>>) {
     super(options);
 
-    this.sourceEvent?.on(async (arg, sender) => {
-      const result = await this.emitInt(arg, sender);
+    this.sourceEvent?.on(async (arg) => {
+      const result = await this.emitInt(arg);
       if (result.length) {
         throw new AggregateError(result);
       }
     });
   }
 
-  async emit(arg: TArg, sender: TSender): Promise<unknown[]> {
+  async emit(arg: TArg): Promise<unknown[]> {
     if (this.isReadOnly) {
       throw new Error('Event is readonly');
     }
 
-    return this.emitInt(arg, sender);
+    return this.emitInt(arg);
   }
 
-  protected async emitInt(arg: TArg, sender: TSender): Promise<unknown[]> {
-    if (this.filter && !this.filter(arg, sender)) {
+  protected async emitInt(arg: TArg): Promise<unknown[]> {
+    if (this.filter && !this.filter(arg)) {
       return [];
     }
 
@@ -36,7 +32,7 @@ export class AsyncEvent<TArg, TSender = void> extends BaseEvent<
 
     for (const fn of this.subscribers) {
       try {
-        await fn.call(sender, arg, sender);
+        await fn(arg);
       } catch (e) {
         errors.push(e);
         if (this.isBailMode) {
@@ -48,7 +44,7 @@ export class AsyncEvent<TArg, TSender = void> extends BaseEvent<
     return errors;
   }
 
-  for(filter: Filter<TArg, TSender>): AsyncEvent<TArg, TSender> {
-    return new AsyncEvent<TArg, TSender>({ bail: this.isBailMode, filter, source: this });
+  for(filter: Filter<TArg>): AsyncEvent<TArg> {
+    return new AsyncEvent<TArg>({ bail: this.isBailMode, filter, source: this });
   }
 }